Transient case for 2 Schottky and simulation of DLTS
Small changes to the strucrute: this time the Schottky barrier is on the left and right side of a domain.
I also added the comments to the file "estimation of time"
The new folder was created to simulate DLTS experiment where the voltage was added to tha sampe e.g. from 0V do -0.5V (so called DLTS signal), form (-0.5V to 0V). One need to remember that the end state of DLTS signal is a starting state of RDLTS signal